Cleanup and fix some issues #101
Ok This will work but you need to change the project layout which I have done. So the Idea here was to move the macros os checks for excludes out of the package def. 
Then wrap the package in an os check macro that throws an error when used with an unsupported os.

I also added in a providers array to apt-get the Linux package  "libv4l-dev".

Now I got this working but I created a new repo with just the framework folder. Also, I had to add a script as you can not have a mix source package so the Linux folder (module) has c and swift.

Now Don't merge this just yet as the project layout need changing to let this be a Swift Package that swift package manager to install.

So if you message me back on this pull request that you would be happy with a project layout change then I will add that to this pull.
Or the other option is to create a 2nd repo with the contents of the Framework folder and this package.swift file and that can be the repo that swift package manager can work with.